backend/pyproject.toml
[project]
name = "helldivers"
version = "0.1.0"
description = ""
authors = [
    {name = "GoodGD",email = "dionatha.work@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-allauth (>=65.12.0,<66.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=46.0.2,<47.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "dj-rest-auth[with-social] (>=7.0.1,<8.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "django-filter (>=25.2,<26.0)",
    "pillow (>=12.0.0,<13.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
